Complete YourNews separate component
